pyproject.toml
[tool.poetry]
name = "nasomedia"
version = "0.1.0"
description = ""
authors = ["Admin <admin@nasomedia.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.11.8"
fastapi = "^0.54.2"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = "^4.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.8.2"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.3"
psycopg2-binary = "^2.8.5"
alembic = "^1.6.5"
sqlalchemy = "^1.4.23"
pytest = "^5.4.3"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
python-dotenv = "^0.15.0"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
azure-storage-blob = "^12.8.1"
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.8.4"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

